Python/requirements.txt
flask
flask-cors
azure-storage-blob
